+++ card-capture/src/components/common/NavigationBar/NavigationBar.tsx
@@ -36,31 +36,11 @@ const NavigationBar = ({ isTransparent = false }: NavigationBarProps) => {
     };
   }, [isTransparent]);
 
-  /**
-   * 반응형을 위해 window.innerWidth를 저장하여 모바일 화면에서는 버튼들이 숨겨지도록 처리
-   */
-  const [windowWidth, setWindowWidth] = useState(window.innerWidth);
-
-  useEffect(() => {
-    const resizeHandler = () => {
-      setWindowWidth(window.innerWidth);
-    };
-
-    window.addEventListener('resize', resizeHandler);
-
-    return () => {
-      window.removeEventListener('resize', resizeHandler);
-    };
-  }, []);
-
-  // 너비가 640px 이하일 때 요소들을 숨김
-  const isMobile = windowWidth <= 640;
-
   return (
     <div
       className={`fixed left-0 top-0 z-20 flex min-h-[60px] w-full items-center justify-between border-b-[1px] border-border px-[30px] ${isTransparent ? navBg : 'bg-white'}`}
     >
-      {/*{!isMobile && <BeforeLoginNav />}*/}
+      {/*<BeforeLoginNav />*/}
       <AfterLoginNav />
     </div>
   );
